requirements.txt
aiohttp
python-dotenv
websockets